negotiator/pyproject.toml
[project]
name = "negotiator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "a2a-sdk[http-server]>=0.3.12",
    "anthropic>=0.73.0",
    "fastapi>=0.121.2",
    "httpx>=0.28.1",
    "pydantic>=2.12.4",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "sse-starlette>=3.0.3",
    "supabase>=2.24.0",
    "uvicorn[standard]>=0.38.0",
]
